fixed compiler errors/warnings when using recent esp SDK

+++ .pio/libdeps/esp12e/EspSoftwareSerial_ID168/src/SoftwareSerial.h
@@ -152,9 +152,9 @@ public:
         return (0x9669 >> byte) & 1;
     }
     /// The read(buffer, size) functions are non-blocking, the same as readBytes but without timeout
-    size_t read(uint8_t* buffer, size_t size);
+    int read(uint8_t* buffer, size_t size);
     /// The read(buffer, size) functions are non-blocking, the same as readBytes but without timeout
-    size_t read(char* buffer, size_t size) {
+    int read(char* buffer, size_t size) {
         return read(reinterpret_cast<uint8_t*>(buffer), size);
     }
     /// @returns The number of bytes read into buffer, up to size. Times out if the limit set through
